Update Lab 2
